fix RSS not showing in prod
