+++ src/app.html
@@ -5,7 +5,7 @@
 		<link rel="icon" href="%sveltekit.assets%/favicon.ico" />
 		<meta name="viewport" content="width=device-width" />
 		<meta property="og:image" content="/imgs/og-img.png" />
-		<link rel="alternate" type="application/rss+xml" title="Hyprland News" href="/rss" />
+		<link rel="alternate" type="application/rss+xml" title="Hyprland News" href="/rss.xml" />
 		<style></style>
 		%sveltekit.head%
 	</head>
